# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  piemdm_internal_model.Approval:
    properties:
      approvalDefCode:
        description: 审批定义编码
        maxLength: 128
        type: string
      code:
        description: 审批实例唯一编码
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 审计字段
        type: string
      currentTaskID:
        description: 当前任务ID
        maxLength: 128
        type: string
      currentTaskName:
        description: 当前任务名称
        maxLength: 128
        type: string
      deletedAt:
        allOf:
        - $ref: '#/definitions/gorm.DeletedAt'
        description: 删除时间
      description:
        description: 申请说明
        type: string
      entityCode:
        description: 关联实体编码
        maxLength: 64
        type: string
      formData:
        description: 表单数据
        type: string
      formSchema:
        description: 表单结构JSON
        type: string
      id:
        type: integer
      priority:
        description: 审批信息
        type: integer
      serialNumber:
        description: 流程信息
        maxLength: 128
        type: string
      status:
        description: |-
          状态信息，审批实例状态
          可选值有：
          - Pending：审批中
          - Approved：通过
          - Rejected：拒绝
          - Canceled：撤回
          - Deleted：删除
        type: string
      title:
        description: 审批标题
        maxLength: 128
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        description: 更新人
        type: string
      urgency:
        description: 紧急程度：Low/Normal/High/Urgent
        type: string
    required:
    - approvalDefCode
    - title
    type: object
  piemdm_internal_model.ApprovalDefinition:
    properties:
      approvalSystem:
        description: |-
          状态和元数据，
          SystemBuilt 系统内置，Custom 自定义，Feishu，DingDing，WeChat
        type: string
      code:
        description: 审批定义唯一编码
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 审计字段
        type: string
      deletedAt:
        allOf:
        - $ref: '#/definitions/gorm.DeletedAt'
        description: 删除时间
      description:
        description: 审批描述
        maxLength: 500
        type: string
      formData:
        description: |-
          表单配置
          FormConfig string `gorm:"type:text"` // 表单配置JSON
          FormSchema string `gorm:"type:text"` // 表单结构定义
        type: string
      id:
        type: integer
      name:
        description: 审批名称
        maxLength: 128
        type: string
      nodeList:
        description: |-
          流程配置
          ProcessConfig string `gorm:"type:text"` // 流程配置JSON
          ProcessNodes  string `gorm:"type:text"` // 流程节点配置
        type: string
      status:
        description: 状态:Normal 正常, Frozen 已冻结, Deleted 已删除
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        description: 更新人
        type: string
    required:
    - name
    type: object
...
host: localhost:8787
info:
  contact:
    email: support@piemdm.com
    name: API Support
  description: PieMDM Master Data Management System API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PieMDM API
  version: "1.0"
paths:
...
  /api/v1/approval-defs/entity/{entityCode}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 实体编码
        in: path
        name: entityCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
            type: array
      summary: 根据实体编码获取审批定义列表
      tags:
      - 审批定义
...
  /api/v1/approvals/entity/{entityCode}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 实体编码
        in: path
        name: entityCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.Approval'
            type: array
      summary: 根据实体获取审批实例列表
      tags:
      - 审批实例
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"